Add `classes` parameter to form elements
There was no way to modify the `class` attribute
of the form elements, which prevented the use
of `is-valid` and `is-invalid` selectors.

See: https://getbootstrap.com/docs/4.3/components/forms/#server-side

Also removed `classes` from `label` and the
wrapper.

AFAIK, it's not necessary to adjust
`label` classes, and the wrapper will probably
be removed from the macros and instead included in
the calling templates.