Use `_bootstrap` for namespace name
While Pattern Lab 3.0 nativeliy allows Twig namespace definition in the
config file (`patternlab-config.json`), the PHP version appears to
require the `evanlovely/plugin-twig-namespaces` and configuration in
`config.yml`.

Yet, I haven't found a way to include a preset `config.yml` with the
starterkit. Without that, it's impossible to predefine a Twig namespace,
even when `evanlovely/plugin-twig-namespaces` is a composer dependency.

I want to keep the underscore in the directory name (`_bootstrap`) so
that the directory will automatically be hidden in `3.x`.

Without `evanlovely/plugin-twig-namespaces` to explicitly define a
namespace of `bootstrap`, Pattern Lab only recognizes the
automatically-detected namespance of `_bootstrap`.

So, I've changed the templates to use `_bootstrap` as the namespace.
Hopefully this will bridge the gap with least friction.